refactor: MonsterEntityWriter を MonsterWriter に改名
前コミットで MonsterEntity クラスを削除したため、セーブ用のライター
クラス名 MonsterEntityWriter が存在しない型名を引きずる形になっていた。
命名を実態に合わせて整理する。

変更:
- クラス名 MonsterEntityWriter → MonsterWriter
- ファイル名 monster-entity-writer.{h,cpp} → monster-writer.{h,cpp}
- src/save/floor-writer.cpp の include / 呼出も追従
- src/Makefile.am のビルド対象リスト更新
- VisualStudio/Bakabakaband/Bakabakaband.vcxproj{,.filters} も更新

既存の命名規則 (lore-writer / player-writer / floor-writer) に整合。
実体はすでに const CreatureEntity &monster を引数に取る形で統合済み。
