+++ VisualStudio/Bakabakaband/Bakabakaband.vcxproj.filters
@@ -2471,7 +2471,7 @@
     <ClCompile Include="..\..\src\monster\monster-timed-effects.cpp">
       <Filter>monster</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\save\monster-entity-writer.cpp">
+    <ClCompile Include="..\..\src\save\monster-writer.cpp">
       <Filter>save</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\timed-effect\player-protection.cpp">
@@ -5700,7 +5700,7 @@
     <ClInclude Include="..\..\src\market\melee-arena.h">
       <Filter>market</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\save\monster-entity-writer.h">
+    <ClInclude Include="..\..\src\save\monster-writer.h">
       <Filter>save</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\system\enums\grid-flow.h">
