+++ src/Makefile.am
@@ -1235,7 +1235,7 @@ hengband_SOURCES = \
 	save/info-writer.cpp save/info-writer.h \
 	save/item-writer.cpp save/item-writer.h \
 	save/lore-writer.cpp save/lore-writer.h \
-	save/monster-entity-writer.cpp save/monster-entity-writer.h \
+	save/monster-writer.cpp save/monster-writer.h \
 	save/player-class-specific-data-writer.cpp save/player-class-specific-data-writer.h \
 	save/player-writer.cpp save/player-writer.h \
 	save/save.cpp save/save.h \
